Display count of cards already added this day
Users will be reminded of the number of cards they loaded on this day
when given the choice to load a new card. This is to avoid people
loading too many cards each day, as that would make the learning process
more difficult. In the future, a warning message might be added when
more than X cards have already been loaded the same day.